[SHARED] Rename common folder fs => findRootPath
