+++ packages/common/src/findRootPath/index.js
